backend/requirements.txt
# Dependencias principales para Mediaris 2.0 con Google Gemini
Flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
google-generativeai==0.8.1
